Add the web page after changing the host to sourceforge.
This also provides a Makefile to build the web version of the
man pages which should work at sourceforge. These will have
to be copied to the /home/project-web/pam-abl/htdocs/docs
directory.  In fact, the whole web directory will have to be
copied or symlinked (if that is even possible in their setup).