BUG3564436 second authentication succeeds
When an application (sshd for example) wants to authenticate a user, it
opens a pam_handle and calls authenticate. If pam_abl is needed, we are
loaded into memory and we are asked to authenticate the user.
In that case we open our db environment and check if the user/host is
blocked or not.
When authentication finishes, we are asked to cleanup, at which point we
look at the authentication result and log it into the db if needed.
Apparently sshd uses the same pam_handle between the attempts, in which
case we are asked multiple times to authenticate a user. Resulting in us
trying to open the same db environment multiple times (that's when you
see the "The database environment could not be opened" message). Because
cleanup is only called when all the attempts have been tried, we only
log one failure instead of three.

The fix was to see if we are called multiple times or not, and if so log
previous calls as failed attempts.