security fix, do not run commands using system()
If the user specifies a clear/block command an attacker can execute
shell code. This is because we use the system() command to execute the
resulting command and this without even trying te escape shell
characters.

So instead of calling system(), we now execute the command using a
fork/exec solution. We also introduced a syntax (using [] ) to specify
the command and it's arguments. This way we are more flexible then just
splitting the command on spaces.
Because this syntax can cause some problems, we also introduced the -d
option, that prints the parsed command.

NOTE: the user can still cause a security problem if he specifies a
shell as it's command. That is because the user/host/service are not
escaped when substituted.