Database module code compiles.
This strips all database related code out of the all the pam and command
line tool functions, replacing it with calls to function pointers which
handle their own state.  The idea is that the database just gets told
what to do, and does the right thing.  This seems better than having every function
have to know how to open a particular database and handle transactions
etc.  It also makes it easier to add different databases or even
different versions of the same one in the future.