Make args global
Since everything needs the arg structure, stop passing it between
everything as an argument, and just make it global.

If it *really* needs to be locked down with something more fancy than a
mutable global because pam_abl's awesomeness makes it so popular and the
code base so huge that the global becomes unmaintable with hundreds of
developers making changes willy nilly, then I suppose there will have to
be a different design (and not the one I just got rid of).  As it was,
I was annoyed with it appearing in every single function definition.  It's
pam_abl.  Anything in pam_abl most probably needs the configuration
for pam_abl.