Fix test.c command line args
I had introduced the need to specify a database module, while danta had
introduced the need to understand -e as the first argument.  This works
now, and the program can differentiate when it just needs to run danta's exit
test and when it needs to start all the tests.  I updated the comment as
well to make it clear what was happening there.