Close db between checks to avoid locking.
Since we no longer need to worry about the cost of opening the database
with Kyoto Cabinet, we can close the database immediately when we are done with
it, and open it later for the next check.  This is more in line with the
way anything using the command line tool would work anyway.

This should also give better concurrency by reducing contention when
handling slow authentication like ssh passwords etc.