do not force the user to have both bdb and kc
Change the build files to only build the db libraries that are actually
present on the system. So if you don't have Berkeley db, we will not try
to make the library for Berkeley db.

Also added the same override flags for Kyoto Cabinet as we had for
Berkeley DB. This way you can link to any version you want and are not
forced to use the system default.