BUG #33 test.c fails to compile on hurd-i386
thanx to  Alex Mestiashvili for reporting this bug.

In function 'removeDir':
error: 'PATH_MAX' undeclared (first use in this function)

I looked at the posix standard. This is what I found:
According to POSIX.1-2001 a buffer of size PATH_MAX suffices, but
PATH_MAX need not be a defined constant and may have to be obtained
using pathconf(3). Asking pathconf(3) does not really help, since, on
the one hand POSIX warns that the result of pathconf(3) may be huge and
unsuitable for mallocing memory, and on the other hand pathconf(3) may
return -1 to signify that PATH_MAX is not bounded.

So just to be sure, I have replaced the piece with a dynamic buffer
just to be sure.