introduce the old pam config behaviour again
The 'new' way of using pam-abl may scare a lot of novice users. The code
to reintroduce the previous behaviour is not that hard and has only
impact on the pam interface, see pam_functions.c

I have done minimal testing. Next thing is to seperate the pam calls
from the pam-abl code, so we can do proper integration tests